+++ src/hapi/pipelines/database/conflict_event.py
@@ -8,6 +8,7 @@ from hapi_schema.utils.enums import EventType
 from hdx.utilities.dateparse import parse_date_range
 from sqlalchemy.orm import Session
 
+from ..utilities.batch_populate import batch_populate
 from ..utilities.logging_helpers import add_message
 from . import admins
 from .base_uploader import BaseUploader
@@ -36,7 +37,7 @@ class ConflictEvent(BaseUploader):
         errors = set()
         for dataset in self._results.values():
             dataset_name = dataset["hdx_stub"]
-            rows = []
+            conflict_event_rows = []
             number_duplicates = 0
             for admin_level, admin_results in dataset["results"].items():
                 # TODO: this is only one resource id, but three resources are downloaded per dataset
@@ -46,6 +47,7 @@ class ConflictEvent(BaseUploader):
                 values = admin_results["values"]
 
                 for admin_code in admin_codes:
+                    admin_rows = []
                     admin2_code = admins.get_admin2_code_based_on_level(
                         admin_code=admin_code, admin_level=admin_level
                     )
@@ -71,20 +73,7 @@ class ConflictEvent(BaseUploader):
                             time_period_range = parse_date_range(
                                 f"{month} {year}", "%B %Y"
                             )
-                            row = (
-                                resource_id,
-                                admin2_code,
-                                event_type,
-                                events,
-                                fatalities,
-                                time_period_range[0],
-                                time_period_range[1],
-                            )
-                            if row in rows:
-                                number_duplicates += 1
-                                continue
-                            rows.append(row)
-                            conflict_event_row = DBConflictEvent(
+                            conflict_event_row = dict(
                                 resource_hdx_id=resource_id,
                                 admin2_ref=self._admins.admin2_data[
                                     admin2_code
@@ -95,12 +84,20 @@ class ConflictEvent(BaseUploader):
                                 reference_period_start=time_period_range[0],
                                 reference_period_end=time_period_range[1],
                             )
-                            self._session.add(conflict_event_row)
-                    self._session.commit()
+                            if conflict_event_row in admin_rows:
+                                number_duplicates += 1
+                                continue
+                            admin_rows.append(conflict_event_row)
+                            conflict_event_rows.append(conflict_event_row)
+
             if number_duplicates > 0:
                 add_message(
                     errors, dataset_name, f"{number_duplicates} duplicate rows"
                 )
+            if len(conflict_event_rows) == 0:
+                add_message(errors, dataset_name, "no rows found")
+                continue
+            batch_populate(conflict_event_rows, self._session, DBConflictEvent)
 
         for dataset, msg in self._config.get(
             "conflict_event_error_messages", dict()